requirements.txt
Django==4.1.5
environs==9.5.0
djangorestframework==3.14.0
requests==2.28.2
redis==4.4.2
